Cant save icon if none exists

+++ app/src/main.js
@@ -23,18 +23,26 @@ app.on('ready', () => {
   // Renderer Listeners
   ipcMain.on('save', (event, arg) => {
     const svg = arg
-    // Save pop up
-    dialog.showSaveDialog((fileName) => {
-      // Write to SVG
-      fs.writeFileSync(fileName + '.svg', svg)
-      // Write to PNG fom SVG
-      gm(fileName + '.svg')
-      .write(fileName + '.png', (err) => {
-        if (err) throw (err)
-        // Delete the SVG
-        fs.unlink(fileName + '.svg')
+    if (svg !== '') {
+      // Save pop up
+      dialog.showSaveDialog((fileName) => {
+        // Write to SVG
+        fs.writeFileSync(fileName + '.svg', svg)
+        // Write to PNG fom SVG
+        gm(fileName + '.svg')
+        .write(fileName + '.png', (err) => {
+          if (err) throw (err)
+          // Delete the SVG
+          fs.unlink(fileName + '.svg')
+        })
       })
-    })
+    } else {
+      dialog.showMessageBox({
+        message: 'No Icon',
+        detail: 'Add an icon by selecting one from the dropdown menu',
+        buttons: ['Ok']
+      })
+    }
   })
   ipcMain.on('about', () => {
     const date = new Date()
